feat(authorization): add Gates and model-based Policies

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -17,7 +18,8 @@ class AttendeeController extends Controller
 
     public function index(Event $event)
     {
-       
+        Gate::authorize('viewAny', $event);
+
         $attendees = $this->loadRelationships(
             $event->attendees()->latest()
         );
@@ -29,6 +31,8 @@ class AttendeeController extends Controller
 
     public function store(Request $request, Event $event)
     {
+        Gate::authorize('create', Attendee::class);
+
         $attendee = $this->loadRelationships(
             $event->attendees()->create([
                 'user_id' => $request->user()->id
@@ -43,6 +47,7 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
+        Gate::authorize('view', $attendee);
         return new AttendeeResource(
             $this->loadRelationships($attendee)
         );
@@ -53,6 +58,8 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete', $attendee);
+        
         $attendee->delete();
 
         return response(status: 204);
